refactor: move CSV upload logic to service
